requirements.txt
wheel==0.38.4
django==4.1.6
django-cors-headers==3.11.0
djangorestframework==3.14.0
mssql-django==1.2
pyodbc==4.0.35
python-dotenv==0.21.1
django-environ==0.9.0